+++ app/src/main/java/app/revanced/manager/util/FileFilters.kt
@@ -9,7 +9,7 @@ fun isAllowedApkFile(path: Path): Boolean {
     return extension in APK_FILE_EXTENSIONS
 }
 
-fun isAllowedRvpFile(path: Path): Boolean {
+fun isAllowedPatchBundleFile(path: Path): Boolean {
     val extension = path.name.substringAfterLast('.', "").lowercase(Locale.ROOT)
-    return extension == "rvp"
+    return extension == "rvp" || extension == "mpp"
 }